+++ src/App.css
@@ -192,3 +192,31 @@ button.mad:hover {
     filter: none !important;
   }
 }
+
+/* InfoPopup.tsx */
+.modal {
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+  flex-direction: column;
+  padding: 1em;
+  
+  filter: drop-shadow(0 0 0.5em #242424);
+  
+  background-color: #e3e3e3;
+  color:#242424;
+  border: 2px solid rgb(240, 240, 240);
+  border-radius: 12px;
+  position:  absolute;
+  width: 40em;
+  /* subtract half modal width, padding, border */
+  left: calc(50% - 20em - 1em - 2px);
+}
+
+@media (width <= 800px) {
+  .modal {
+    width: 90vw;
+    /* subtract half modal width, padding, border */
+    left: calc(50% - 45vw - 1em - 2px);
+  }
+}
